Remove postinstall script altogether
After a couple iterations, I've decided the best approach to this
is to make sure the project is never built when installed into a
project.

The previous workaround (#83 + #84) was insufficient because
while it dropped the requirement of the host project to have
typescript installed, it still required various @types/ packages
to be installed, so the solution was insufficient.

The hope was to preserve behavior that when this project is
cloned and installed, a build would be done. I don't see
a nice way to accomplish this without also having the effect
of running the build script when this package is installed into
another project.

The `prepare` script was considered as well, but that script
is called both for transitive and local installs which is
what we want to avoid.

The `prepack` script is necessary for CI / publish steps to
ensure the build is completed before publishing to NPM.
(a learning from https://github.com/apollographql/datasource-rest/pull/57)